Support style hierarchy and composition in StyleProvider
The OutputElementStyle's values map properties can now override the
properties defined for the style identified by the parent key.

A multi-level hierarchy of parent styles is supported. Style composition
by using a comma-delimited list of parent keys is supported.

The 'background-color', 'height', 'border-style', 'border-color' and
'border-width' properties are supported for time graph states.

The 'symbol-type', 'color', 'height' and 'vertical-align' properties are
supported for time graph annotations.

The 'opacity' property and '-blend' modifier are supported for color
styles.

The '-factor' modifier is supported for number styles.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>