Support tooltip for annotations
Move the fetch tooltip implementation to the TspDataProvider class.
Provide a method to fetch state tooltip and a method to fetch annotation
tooltip. Create the requested element for the request from the element
model.

Add the category to the annotation model when fetching annotations.

Adapt the tooltip content according to the type of element.

Signed-off-by: Patrick Tasse <patrick.tasse@ericsson.com>