Add data tree output component implementation
The implementation is the first iteration. It introduces the component
class which uses the tree implementation also used in the time-graph and
XY components. The component is not enabled in this patch.

The tree implementation need some UI/UX improvement, for example for
styling or column resizing which will be done in a separated commit.

Once some improvements are available, the data tree component
will be shown as part of the available views list for a given trace.

Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>