Display the tooltip for xy
The tooltip offered by ChartJS can be extensively configured, but it
still fails in some edge cases when used in the XY chart.

For instance, it is drawn in the canvas that holds the chart, so it
is cut off when the mouse approaches its borders. Also, when selecting
too many processes, the performance of the app degrades to the point of
making the page unusable.

This commit implements a custom tooltip that does not have the problem
of being cut off at the borders, and performs better under heavy loads.

Main characteristics of the tooltip:

- It summarizes all processes that have value equal to zero in a single line;
- It allows the user to move the mouse into it, to scroll and to select text;
- It is aware of the proximity to the borders of the screen, optimizing its
position to avoid clipping.

Signed-off-by: Ibrahim Fradj <ibrahim.fradj@ericsson.com>
Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>