Remove Theia dependency in the react-components package
Pull Request #556 accidentally introduced a dependency to Theia in the
react-components package which needs to be independent to Theia
so that it can be re-used in other environments.

To achieve the removal of the Theia dependency as well as keep the
behavior introduced by #556 the following changes have been made:
- Rename 'close' trace/experiment methods and signals to use term
'delete' instead. Close in this case actually deletes the
trace/experiment from the trace server and the original names were
misleading
- Introduce EXPERIMENT_CLOSED signal to notify when a experiment tab
is actually closed
- Listen to EXPERIMENT_CLOSED signal in relevant views to update the
views, i.e. clear list of available views or clear experiment selection
when no trace viewer tab is open.
- rename some method and signal handler names for clarity reasons

Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>