electron: bundle Trace Server
Make the required adjustments to bundle the Trace Server on Electron.

Here's how the application will look for the Trace Server path now:

- `browser`:
  - Read `traceViewer.traceServer.path` preference.
  - If `undefined` or empty, read `TRACE_SERVER_PATH` environment variable.
  - If `undefined` or empty, fail.
- `electron`:
  - Read `traceViewer.traceServer.path` preference.
  - If `undefined` or empty, read `TRACE_SERVER_PATH` environment variable.
  - If `undefined` or empty, use the known path to the bundled server.

Cleanup code in adjacent areas to the modifications.

`traceViewer.traceServer.path` default preference is now `undefined`.

Split the implementation of `TraceServerConfigService` between regular
Node and Electron: on the latter the implementation will default to the
bundled Trace Server.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>