Handle error messages in abstract output component
When error messages are handled in each component rather (than by the
parent class), it is easy to forget to handle these errors when
implementing new output components. Handling errors in the parent class
makes these bugs less likely.

"No results" message: Instead of needing to remember to call the
function in each output components the abstract class structure imposes
that every output component implement a check for empty results (and
displaying the message is handled by the parent class).

Analysis failed message: The developper still needs to remember to set
the output status to failed when implementing new output components.
This change does not help with remembering to do that.

Signed-off-by: Erica Bugden <erica.bugden@gmail.com>