Fix use of TimeGraphArrow sourceId and destinationId
The sourceId and destinationId should be the row id of source row and
destination row. But it is used as the index of the source and
destination rows in the full list of rows.

Change to use the sourceId and destinationId properly. This requires
that the TimeGraphChartArrows layer receive the full list of row ids.

Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>