xy-output-component: Deserialize for isScatterPlot
Set the isScatterPlot flag based on the now possibly serialized series
type, as opposed to the early OutputDescriptor ID prior. Still rely on
the latter by default if using a server that doesn't serialize types
yet. Contributes to fixing issue #45. Refactor the otherwise duplicated
xyDataResponse.model.series use.

Assume that type being set through the first series, should there be
many. This is necessary as the chart type is currently for all series,
as opposed to eventually supporting one such type per series. This would
mean supporting a potential mix of scatter and line (or else) series to
draw side-by-side on a single chart.

This change depends on the tsp-typescript-client model change [1] below,
and can be tested manually this way:

1. run theia-trace-extension as instructed in ./README.md;
2. open a trace such as [2];
3. open any one of the Latency vs Time analyses;
4. select any or all available checkboxes to view the said scatter plot.

Update yarn.lock's tsp-typescript-client@next with merged [1]'s version.

[1] XY: Add the series OutputElementStyle to the model
[2] https://git.eclipse.org/r/plugins/gitiles/tracecompass/tracecompass-test-traces/+/refs/heads/master/ctf/src/main/resources/kernel

Signed-off-by: Marco Miller <marco.miller@ericsson.com>