Select chart components when selecting an event with keyboard in table
When using the keyboard to select an event, broadcast the event details
to other charts to select corresponding chart component.

Add support for arrow up/down key selection of events in the events
table.

When extending the selection, only the last selected event will be
used to broadcast the selection (multi-line selection).

Make sure that the properties view is updated with the last event
selection.

Signed-off-by: Bernd Hufmann <bernd.hufmann@ericsson.com>