Fix theia-trace-extension table vertical scrollbar
Before this commit, the size of the table was not set to the  number of
elements displayed in the table so the vertical scrollbar was not
adjusted and didn't reflect the number of elements left in the table.
This commit fixes this issue by setting the size of the table to the
real number of elements so the scrollbars are updated.

Signed-off-by: KyrollosBekhet <kyrollos.bekhet@ericsson.com>