Fix for unresponsive big tables
Because of Ag-grid limitations, traces that have tables with too many
rows (more than 600,000) have the issue of not allowing interaction
with the table when the user is near its bottom.

This solution implements a pagination feature from Ag-grid, which breaks
large tables into pages and eliminates the problem of unresponsiveness.

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>