Improve color scheme for zoom range of overview output
Currently, the zoom range is shown in the trace overview
using an orange overlay. When the users want to select
a range (indicated by a blue overlay), these two layers
overlap and make it hard to distinguish them. This commit
improves the color scheme for the zoom range by 1) remove
the orange overlay of the zoom range and 2) use a grey
overlay to indicate the part of the trace that is not zoomed.

Fixes #857.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>