Fix pagination bar bug
With traces that have very large tables, if the event table is opened
before the trace is fully loaded, the pagination bar does not appear.

The problem is that Ag-grid does not make the pagination bar reactive,
so it only responds to the initial values of the table. Updates of the
number of rows are ignored.

The solution is to create a custom pagination bar that can respond
reactively to data being loaded.

Fixes #849

Signed-off-by: Rodrigo Pinto <rodrigo.pinto@calian.ca>