Time Range Data Widget
Adds the Time Range Data Widget to the sidebar as specified in
ADR 0008: Time Range Data Widget

Time range changes are dispatched via the SignalManager.  These
changes are received by the Time Range Data Widget and stored into a
map of time range data for all experiments.  When the user inputs a
valid selection range value in the widget, the value is dispatched via
SignalManager, received in Trace Context Component, and set to it's
Unit Controller's selection range (only if the experiment id matches).

Signed-off-by: William Yang <william.yang@ericsson.com>