Abstract CSS style for React components
Since the Trace Extension was migrated to VSCode, the CSS classes must be
abstracted and mapped to the correct Theia and VSCode CSS variables. This
commit replaces the prefix --theia of CSS variables with --trace-viewer in
the react-components package. For example, the variable --theia-ui-padding
becomes --trace-viewer-ui-padding. The packages that use these components
then have to import a CSS file that maps these CSS variables to the CSS
variables of their local environment, or they need to define the CSS
for those components.

Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>