Improved UX of column search, results
Upon pressing Enter with search text, a search is triggered. During the
search, a spinning loading icon is displayed in the column next to the
search text input field.

If there are results, the same behaviour as before is used. If no
results match the search the search text goes red. Any changes to the
search text reset the color of the search text.

Modified hover text to tell user to press Enter, added hover text to other icons.

This makes using the column search more clear, and makes the user aware
of the search status and if there are any results.

Signed-off-by: Dylan leclair dleclair@blackberry.com