Add and handle data changed signal
Introduced a new signal to indicate that the data on the server has
been updated and therefore the components should refetch the data and
rerender the components. This PR handles the signal in table-output-component
and timegraph-output-component. We have a use case where data is dynamically
modified on the server, and we want the components to rerender with the new data.
Currently, the user would have to close the active trace panel and reopen it to
see the changes. Added output descriptors as the signal payload.

Signed-off-by: Neel Gondalia <ngondalia@blackberry.com>