Use TimeGraphContainer noWebGl option
There's a relatively low limit of WebGL contexts available for a
Chrome-based browser tab (16). Each time a "TimeGraphContainer" is
created, one such context is consumed. We use these contexts for a
few things, not just time-line charts (where it makes most sense).

This commit takes advantage of the new option and declines that a
WebGL context be used for containers we create, except for proper
chart views. This means that we will now be able to create about
16 timeline-chart views per session, spread over however many
experiments.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>