[traceviewer-libs] Add readme, GitHub workflow and templates,
and expected root folder boilerplate files.

The workflow is a minimal one, that manually builds the two libraries
and runs the unit tests available in traceviewer-react-components.
Since the package.json one level up is not available in the subtree,
we get creative, manually adding the missing dependencies necessary
to sucessfully run the tests.

Everything before this commit is the result of running "git subtree
split" to extract the traceviewer libraries from their original
"home", under root folder "packages" in repo theia-trace-extension.
Full history touching the libraries was preserved, though the commit
messages may refer to things other than the libraries, that may not
be relevant here, since the commits were made in the context of the
original repository.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>